+++ src/main/java/com/genius/herewe/business/location/facade/DefaultLocationFacade.java
@@ -4,8 +4,6 @@ import static com.genius.herewe.core.global.exception.ErrorCode.*;
 
 import java.util.List;
 
-import org.hibernate.StaleStateException;
-import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +12,7 @@ import com.genius.herewe.business.location.LocationRequest;
 import com.genius.herewe.business.location.domain.Location;
 import com.genius.herewe.business.location.dto.LocationInfo;
 import com.genius.herewe.business.location.dto.PlaceResponse;
+import com.genius.herewe.business.location.handler.RetryHandler;
 import com.genius.herewe.business.location.search.dto.Place;
 import com.genius.herewe.business.location.service.LocationService;
 import com.genius.herewe.business.moment.domain.Moment;
@@ -30,15 +29,21 @@ import lombok.extern.slf4j.Slf4j;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class DefaultLocationFacade implements LocationFacade {
-	private final int MAX_RETRIES = 3;
-
 	private final MomentService momentService;
 	private final MomentMemberService momentMemberService;
 	private final LocationService locationService;
+	private final RetryHandler retryHandler;
 
 	@Override
 	@Transactional
 	public Place addPlace(Long momentId, LocationRequest locationRequest) {
+		return retryHandler.executeWithRetry(
+			() -> executeAddPlace(momentId, locationRequest)
+		);
+	}
+
+	@Transactional
+	public Place executeAddPlace(Long momentId, LocationRequest locationRequest) {
 		try {
 			Place place = locationRequest.place();
 
@@ -49,7 +54,7 @@ public class DefaultLocationFacade implements LocationFacade {
 			}
 
 			Moment moment = momentService.findByIdWithOptimisticLock(momentId);
-			
+
 			Location location = Location.createFromPlace(place, nextIndex);
 			location.addMoment(moment);
 			locationService.save(location);
@@ -80,63 +85,10 @@ public class DefaultLocationFacade implements LocationFacade {
 	@Override
 	@Transactional
 	public void deletePlace(Long userId, Long momentId, int locationIndex) {
-		int retryCount = 0;
-		while (retryCount < MAX_RETRIES) {
-			try {
-				// 이 밑에 함수만 다른걸 쓸거임
-				executeDeletePlace(userId, momentId, locationIndex);
-				return;
-			} catch (Exception e) {
-				retryCount++;
-				if (shouldRetry(e) && retryCount < MAX_RETRIES) {
-					logRetryAttempt(retryCount, momentId, locationIndex, e);
-					applyBackoff(retryCount);
-				} else {
-					throw translateException(e);
-				}
-			}
-		}
-	}
-
-	private boolean shouldRetry(Exception e) {
-		boolean isDbException = e instanceof OptimisticLockException
-			|| e instanceof DataAccessException
-			|| e instanceof StaleStateException;
-
-		boolean isRetryException = (e instanceof BusinessException) &&
-			(((BusinessException)e).getErrorCode() == CONCURRENT_MODIFICATION_EXCEPTION);
-
-		boolean hasRetryCause = e.getCause() != null && (
-			e.getCause() instanceof DataAccessException ||
-				(e.getCause().getMessage() != null &&
-					e.getCause().getMessage().contains("Deadlock"))
-		);
-
-		return isDbException || isRetryException || hasRetryCause;
-	}
-
-	private void applyBackoff(int retryCount) {
-		try {
-			long delay = (long)(100 * Math.pow(2, retryCount - 1));
-			Thread.sleep(delay);
-		} catch (InterruptedException ie) {
-			Thread.currentThread().interrupt();
-		}
-	}
-
-	private void logRetryAttempt(int retryCount, Long momentId, int locationIndex, Exception e) {
-		log.warn("Retry attempt {}/{} for deleting place: momentId={}, locationIndex={}, error={}",
-				 retryCount, MAX_RETRIES, momentId, locationIndex, e.getMessage());
-	}
-
-	private RuntimeException translateException(Exception e) {
-		if (e instanceof BusinessException) {
-			return (BusinessException)e;
-		} else if (e instanceof OptimisticLockException || e instanceof DataAccessException) {
-			return new BusinessException(CONCURRENT_MODIFICATION_EXCEPTION, e);
-		} else {
-			return new BusinessException(UNEXPECTED_ERROR, e);
-		}
+		retryHandler.executeWithRetry(() -> {
+			executeDeletePlace(userId, momentId, locationIndex);
+			return null;
+		});
 	}
 
 	@Transactional
