+++ src/main/java/com/genius/herewe/business/location/handler/RetryableOperation.java
@@ -1,0 +1,6 @@
+package com.genius.herewe.business.location.handler;
+
+@FunctionalInterface
+public interface RetryableOperation<T> {
+	T execute() throws Exception;
+}
